requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-admin-interface==0.24.2
django-ckeditor==6.6.1
django-colorfield==0.8.0
django-jalali-date==1.0.1
django-js-asset==2.1.0
django-widget-tweaks==1.4.12
idna==3.4
jdatetime==4.1.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
